+++ lib/service_layer/command.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'service_layer/result'
+
 module ServiceLayer
   # The +Command+ contains the logic related to the command pattern. This
   # pattern is one of the bases of the service layer. It provides the logic
@@ -40,7 +42,9 @@ module ServiceLayer
   #   end
   #
   #   PeriodConverter.perform(period: :this_week)
-  #   # => {:range=>2017-10-01 00:00:00 +0200..2017-10-07 12:17:20 +0200}
+  #   # => #<ServiceLayer::Result:0x00558a4cc55428
+  #        @range=2017-10-01 00:00:00 +0200..2017-10-07 12:17:20 +0200,
+  #        @success=true>
   module Command
     # @!visibility private
     def self.included(base)
@@ -54,8 +58,9 @@ module ServiceLayer
       # Delegates the initialization, and performs the service.
       #
       # @param properties [Hash]
+      # @return [Result]
       def perform(**properties)
-        new(**properties).perform
+        new(**properties).execute
       end
     end
 
@@ -74,7 +79,19 @@ module ServiceLayer
       end
     end
 
+    # Encapsulates the execution of the service and determines the result state.
+    #
+    # @return [Result]
+    def execute
+      result = perform
+    rescue StandardError => exception
+      Result.new(error: exception).tap(&:fail!)
+    else
+      Result.new(result)
+    end
+
     # @abstract must implement perform method when include command pattern.
+    # @return [Hash]
     def perform
       raise NotImplementedError.new('Service must implement a perform method')
     end